Break apart different BrowserFragment versions
