--- app/src/main/java/org/mozilla/reference/browser/CustomTabActivity.kt
@@ -1,7 +1,0 @@
-/* This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
-
-package org.mozilla.reference.browser
-
-class CustomTabActivity : BrowserActivity()
