+++ app/src/main/java/org/mozilla/reference/browser/IntentReceiverActivity.kt
@@ -7,26 +7,30 @@ package org.mozilla.reference.browser
 import android.app.Activity
 import android.content.Intent
 import android.os.Bundle
-import mozilla.components.browser.session.tab.CustomTabConfig
-import mozilla.components.support.utils.SafeIntent
+import kotlinx.coroutines.MainScope
+import kotlinx.coroutines.launch
 import org.mozilla.reference.browser.ext.components
 
 class IntentReceiverActivity : Activity() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-
         val intent = intent?.let { Intent(it) } ?: Intent()
+        val processors = components.utils
 
-        components.utils.intentProcessor.process(intent)
+        MainScope().launch {
+            processors.customTabIntentProcessor.process(intent) ||
+                processors.tabIntentProcessor.process(intent)
 
-        if (CustomTabConfig.isCustomTabIntent(SafeIntent(intent))) {
-            intent.setClassName(applicationContext, CustomTabActivity::class.java.name)
-        } else {
-            intent.setClassName(applicationContext, BrowserActivity::class.java.name)
+            val className = if (processors.customTabIntentProcessor.matches(intent)) {
+                ExternalAppBrowserActivity::class
+            } else {
+                BrowserActivity::class
+            }
+            intent.setClassName(applicationContext, className.java.name)
+
+            startActivity(intent)
+            finish()
         }
-
-        startActivity(intent)
-        finish()
     }
 }
